Split Swiftask.h into its component classes under APIContracts project
Also updated all existing references to Swiftask accordingly

Note: there still exists errors in files; e.g. StorageHandler.h
