+++ SwiftaskGUI.sln
@@ -13,6 +13,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UI", "UI\UI.vcxproj", "{D93
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Logic", "Logic\Logic\Logic.vcxproj", "{1A7DEB8B-9BEF-40EB-9CED-EC4676369E05}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "APIContracts", "APIContracts\APIContracts.vcxproj", "{1AE49B26-3E8B-4AD7-9798-126614368F60}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Win32 = Debug|Win32
@@ -31,6 +33,10 @@ Global
 		{1A7DEB8B-9BEF-40EB-9CED-EC4676369E05}.Debug|Win32.Build.0 = Debug|Win32
 		{1A7DEB8B-9BEF-40EB-9CED-EC4676369E05}.Release|Win32.ActiveCfg = Release|Win32
 		{1A7DEB8B-9BEF-40EB-9CED-EC4676369E05}.Release|Win32.Build.0 = Release|Win32
+		{1AE49B26-3E8B-4AD7-9798-126614368F60}.Debug|Win32.ActiveCfg = Debug|Win32
+		{1AE49B26-3E8B-4AD7-9798-126614368F60}.Debug|Win32.Build.0 = Debug|Win32
+		{1AE49B26-3E8B-4AD7-9798-126614368F60}.Release|Win32.ActiveCfg = Release|Win32
+		{1AE49B26-3E8B-4AD7-9798-126614368F60}.Release|Win32.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
